+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Added
 - Support for minimal web testing (#2)
 - Add Python setup
+- Load actions from site-packages
+
+### Changed
+- Source code restriction in LICENSE
 
 ## 0.1.0 - 2019-07-23
 ### Added